requirements.txt
numpy
pandas
matplotlib
torch
torchvision
ipykernel
Pillow
scikit-learn
requests
seaborn
ipywidgets